Improved MSBuild projects - moved output of MSBuild projects to `~output` - moved `hierarchical_mutex` class to separate header/source files pair
