+++ src/1_thread.vcxproj
@@ -73,19 +73,23 @@
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <LinkIncremental>true</LinkIncremental>
-    <IntDir>tmp\$(ProjectName)\$(Platform)-$(Configuration)\</IntDir>
+    <IntDir>$(SolutionDir)~output\tmp\$(ProjectName)\$(Configuration)-$(Platform)\</IntDir>
+    <OutDir>$(SolutionDir)~output\bin\$(ProjectName)\$(Configuration)-$(Platform)\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <LinkIncremental>true</LinkIncremental>
-    <IntDir>tmp\$(ProjectName)\$(Platform)-$(Configuration)\</IntDir>
+    <IntDir>$(SolutionDir)~output\tmp\$(ProjectName)\$(Configuration)-$(Platform)\</IntDir>
+    <OutDir>$(SolutionDir)~output\bin\$(ProjectName)\$(Configuration)-$(Platform)\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <LinkIncremental>false</LinkIncremental>
-    <IntDir>tmp\$(ProjectName)\$(Platform)-$(Configuration)\</IntDir>
+    <IntDir>$(SolutionDir)~output\tmp\$(ProjectName)\$(Configuration)-$(Platform)\</IntDir>
+    <OutDir>$(SolutionDir)~output\bin\$(ProjectName)\$(Configuration)-$(Platform)\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <LinkIncremental>false</LinkIncremental>
-    <IntDir>tmp\$(ProjectName)\$(Platform)-$(Configuration)\</IntDir>
+    <IntDir>$(SolutionDir)~output\tmp\$(ProjectName)\$(Configuration)-$(Platform)\</IntDir>
+    <OutDir>$(SolutionDir)~output\bin\$(ProjectName)\$(Configuration)-$(Platform)\</OutDir>
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <ClCompile>
