+++ src/2_async.vcxproj
@@ -73,22 +73,26 @@
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <LinkIncremental>true</LinkIncremental>
     <IncludePath>..\benchmark\include;$(IncludePath)</IncludePath>
-    <IntDir>tmp\$(ProjectName)\$(Platform)-$(Configuration)\</IntDir>
+    <IntDir>$(SolutionDir)~output\tmp\$(ProjectName)\$(Configuration)-$(Platform)\</IntDir>
+    <OutDir>$(SolutionDir)~output\bin\$(ProjectName)\$(Configuration)-$(Platform)\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <LinkIncremental>true</LinkIncremental>
     <IncludePath>..\benchmark\include;$(IncludePath)</IncludePath>
-    <IntDir>tmp\$(ProjectName)\$(Platform)-$(Configuration)\</IntDir>
+    <IntDir>$(SolutionDir)~output\tmp\$(ProjectName)\$(Configuration)-$(Platform)\</IntDir>
+    <OutDir>$(SolutionDir)~output\bin\$(ProjectName)\$(Configuration)-$(Platform)\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <LinkIncremental>false</LinkIncremental>
     <IncludePath>..\benchmark\include;$(IncludePath)</IncludePath>
-    <IntDir>tmp\$(ProjectName)\$(Platform)-$(Configuration)\</IntDir>
+    <IntDir>$(SolutionDir)~output\tmp\$(ProjectName)\$(Configuration)-$(Platform)\</IntDir>
+    <OutDir>$(SolutionDir)~output\bin\$(ProjectName)\$(Configuration)-$(Platform)\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <LinkIncremental>false</LinkIncremental>
     <IncludePath>..\benchmark\include;$(IncludePath)</IncludePath>
-    <IntDir>tmp\$(ProjectName)\$(Platform)-$(Configuration)\</IntDir>
+    <IntDir>$(SolutionDir)~output\tmp\$(ProjectName)\$(Configuration)-$(Platform)\</IntDir>
+    <OutDir>$(SolutionDir)~output\bin\$(ProjectName)\$(Configuration)-$(Platform)\</OutDir>
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <ClCompile>
