+++ src/6_deadlock_detection.cpp
@@ -1,3 +1,4 @@
+#include "hierarchical_mutex.h"
 #include <atomic>
 #include <chrono>
 #include <cstdint>
@@ -45,77 +46,6 @@ milliseconds RandomTime(milliseconds from, milliseconds to)
 
 	return milliseconds(distribution(generator));
 }
-
-// Класс hierarchical_mutex представляет mutex с обнаружением deadlock.
-// Объекты типа hierarchical_mutex совместно выстраивают иерархию использующих их потоков.
-// Для хранения индекса потока в иерархии используется thread_local переменная.
-// Начальный индекс потока в иерархии - максимальное число unsigned.
-// Затем, при захвате mutex индекс потока становится равным константному индексу mutex,
-//  а mutex запоминает предыдущее значение индекса иерархии потока.
-// Затем, при освобождении mutex индекс потока становится прежним.
-// Если при захвате мьютекса индекс потока (т.е. индекс ранее захватенного mutex)
-//  меньше индекса захватываемого mutex, то произошёл deadlock.
-class hierarchical_mutex
-{
-public:
-	explicit hierarchical_mutex(unsigned long value)
-		: hierarchy_value(value)
-		, previous_hierarchy_value(0)
-	{
-	}
-
-	void lock()
-	{
-		check_for_hierarchy_violation();
-		internal_mutex.lock();
-		update_hierarchy_value();
-	}
-
-	void unlock()
-	{
-		this_thread_hierarchy_value = previous_hierarchy_value;
-		internal_mutex.unlock();
-	}
-
-	bool try_lock()
-	{
-		check_for_hierarchy_violation();
-		if (!internal_mutex.try_lock())
-		{
-			return false;
-		}
-
-		update_hierarchy_value();
-		return true;
-	}
-
-private:
-	std::mutex internal_mutex;
-	unsigned long const hierarchy_value;
-	unsigned previous_hierarchy_value;
-
-	static thread_local unsigned long this_thread_hierarchy_value;
-
-	void check_for_hierarchy_violation()
-	{
-		if (this_thread_hierarchy_value <= hierarchy_value)
-		{
-#if 1 // Текущая стратегия обработки deadlock: вызов abort
-			abort();
-#else
-			throw std::logic_error("mutex hierarchy violated");
-#endif
-		}
-	}
-
-	void update_hierarchy_value()
-	{
-		previous_hierarchy_value = this_thread_hierarchy_value;
-		this_thread_hierarchy_value = hierarchy_value;
-	}
-};
-
-thread_local unsigned long hierarchical_mutex::this_thread_hierarchy_value{ (std::numeric_limits<unsigned long>::max)() };
 
 class Logger
 {
