+++ src/9_thread_pool.vcxproj
@@ -72,23 +72,27 @@
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <LinkIncremental>false</LinkIncremental>
-    <IntDir>tmp\$(ProjectName)\$(Platform)-$(Configuration)\</IntDir>
+    <IntDir>$(SolutionDir)~output\tmp\$(ProjectName)\$(Configuration)-$(Platform)\</IntDir>
     <IncludePath>..\benchmark\include;$(IncludePath)</IncludePath>
+    <OutDir>$(SolutionDir)~output\bin\$(ProjectName)\$(Configuration)-$(Platform)\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <LinkIncremental>true</LinkIncremental>
-    <IntDir>tmp\$(ProjectName)\$(Platform)-$(Configuration)\</IntDir>
+    <IntDir>$(SolutionDir)~output\tmp\$(ProjectName)\$(Configuration)-$(Platform)\</IntDir>
     <IncludePath>..\benchmark\include;$(IncludePath)</IncludePath>
+    <OutDir>$(SolutionDir)~output\bin\$(ProjectName)\$(Configuration)-$(Platform)\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <LinkIncremental>true</LinkIncremental>
-    <IntDir>tmp\$(ProjectName)\$(Platform)-$(Configuration)\</IntDir>
+    <IntDir>$(SolutionDir)~output\tmp\$(ProjectName)\$(Configuration)-$(Platform)\</IntDir>
     <IncludePath>..\benchmark\include;$(IncludePath)</IncludePath>
+    <OutDir>$(SolutionDir)~output\bin\$(ProjectName)\$(Configuration)-$(Platform)\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <LinkIncremental>false</LinkIncremental>
-    <IntDir>tmp\$(ProjectName)\$(Platform)-$(Configuration)\</IntDir>
+    <IntDir>$(SolutionDir)~output\tmp\$(ProjectName)\$(Configuration)-$(Platform)\</IntDir>
     <IncludePath>..\benchmark\include;$(IncludePath)</IncludePath>
+    <OutDir>$(SolutionDir)~output\bin\$(ProjectName)\$(Configuration)-$(Platform)\</OutDir>
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <ClCompile>
